return

+++ MTP_Port2/Float2D.cpp
@@ -1,11 +1,11 @@
 #include "Float2D.h"
 
-Float2D::Float2D(vector<vector<float>>& fVec): row(fVec.size()), col(fVec[0].size())
+Float2D::Float2D(vector<vector<double>>& fVec): row(fVec.size()), col(fVec[0].size())
 {
-	fMatrix = new float*[row];
+	fMatrix = new double*[row];
 	for (size_t i = 0; i < row; i++)
 	{
-		fMatrix[i] = new float[col];
+		fMatrix[i] = new double[col];
 		for (size_t j = 0; j < col; j++)
 		{
 			fMatrix[i][j] = fVec[i][j];
@@ -20,7 +20,7 @@ Float2D::~Float2D()
 	del2D(fMatrix, row);
 }
 
-float ** Float2D::get()
+double ** Float2D::get()
 {
 	return fMatrix;
 }
